requirements.txt
aniso8601==9.0.1
astroid==2.9.3
black==22.1.0
click==8.0.3
colorama==0.4.4
Flask==2.0.2
Flask-Cors==3.0.10
flask-marshmallow==0.14.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn==20.1.0
isort==5.10.1
itsdangerous==2.0.1
Jinja2==3.0.3
lazy-object-proxy==1.7.1
MarkupSafe==2.0.1
marshmallow==3.14.1
marshmallow-sqlalchemy==0.27.0
mccabe==0.6.1
mypy-extensions==0.4.3
mysqlclient==2.1.0
numpy==1.22.2
pandas==1.4.1
pathspec==0.9.0
permissive-dict==1.0.4
platformdirs==2.5.0
pylint==2.12.2
python-dateutil==2.8.2
python-dotenv==0.19.2
python-slugify==6.0.1
pytz==2021.3
six==1.16.0
SQLAlchemy==1.4.31
text-unidecode==1.3
toml==0.10.2
tomli==2.0.1
typing_extensions==4.0.1
waitress==2.0.0
Werkzeug==2.0.3
wrapt==1.13.3
